CRM-17256 fix for paylater billing_required
This commit sets up pay-later as a payment processor class in limited capacity to allow it to cope
with the field loading for billing_required_for_pay_later contribution forms. I had hoped not to have to do this
in 4.7 but the address fields were not loading correctly on form-change :-(. This did allow me to remove quite a few
hacks that were in place to support the way the billing_for_pay_later was working though :-).

Note that a better solution than a check-box for billing-required-for-pay-later would be to allow joining a profile
of field that the paymnet processor should be 'encouraged' to provide  on the payment form. ie. the payment processor could
then build the form to include the profile fields and those required (if any) by it's processor.

The code is written on the assumption the billing profile might be changed to a real profile at some point