CRM-13803 - hook_navigationMenu - Infer navID/parentID
This topic has previously been discussed a bit in
https://issues.civicrm.org/jira/browse/CRM-13803 and
https://github.com/civicrm/civihr/pull/190 .  Long-story short: when an
extension adds a new nav item, it must compute navID and parentID, and many
extensions do this a buggy way.

With this revision, an extension can simply omit the navID and parentID.
The navID will be computed (as the next available ID), and the parentID will
be inferred (based on the navID in the actual parent).