Switch back to retrieving memberships based on membership payment rather than accurate line items
Per comments the unit tests are not building membership payments with accurate linkage between the membership & contribution. THis is not a new
regression so not addressing at the moment. It may related to CRM-14918. NB it seems that the intent is there should be a /bin/bash contribution for a /bin/bash membership
so I guess the line item SHOULD link the 2