uf-field tidyup: deprectate option for autoweight
The autoweight doesn't seem to add any value. If weight is not set on a new entry we should calculate it. If weight have been changed we should calculate the others

However, the ability to opt out is currently leveraged in the replace
function so we need to continue to respect that