Net_SMTP, Auth_SASL, Net_Socket - Load via composer
Note: `pear.php.net` is in the middle of a multiday outage, so downloading
PEAR packages is a bit flaky.  For `Net_Socket` and `Auth_SASL`, we
work-around the outage by referencing specific git commits in
`composer.json`.  Fortunately, these packages are small and don't change
much, so these should be pretty close to the latest stable.