CRM-17718 add Financial type to recur form
This has only been added for those with sites with no ACLs as the financial acl code is quite new & raised questions when I looked.

Further the option is only editable if there is a single line item. Multiple line item recurrings cannot be overridden from here

Contribution Recur Edit should not require installments

Leaving installments empty is valid - ie. it could be null for unlimited

(there is an implication that it should be '0' rather than null but null 'works' & is set in some cases
and we permit the DB field to be null - which we should not do if null is not OK)