CRM_Utils_SQL_Select - Fix dedupe behavior for `select`, `groupBy`, `having`, `orderBy`.
The selected columns should not be deduped at all -- when using `insertInto()`, the selected
columns need to line up with the insertion columns, and it's perfectly reasonably to re-use
the same value in multipl columns.

For th others, it's semantically OK to dedupe, but we should dedupe based on
the evaluated query expression rather than the general formula.