CRM-17837 Report class pager function, test/ api access fix.
This minor patch just checks if the get function is callable before calling it. It allows the api calls to not fatal.
The api calls are primarily used in the unit tests at this stage although the original intent was about more flexible charting by retrieving
the data from an api & passing it to 'any' charting library

Change-Id: I8b3e4e9ade3f75ce8bc6e608cb3c7c6e60d127b9