tools/scripts/phpunit - Delegate to `phpunit` from `PATH`
Embedding a copy of PHPUnit within CiviCRM makes it really confusing when testing
extensions. For example, if an extension author tries to use `phpunit` directly, he'll
conflicts between his copy and the embedded copy.

The work-around has been to *always* use the embedded copy, but this is still confusing
because bootstrapping becomes complicated.

The goal here is to remove Civi's embedded copy of PHPUnit to avoid conflicts.

To preserve compatibility with existing test-documentation and test-scripts
(Jenkins), we leave behind a stub script in `tools/scripts/phpunit`.