CRM-17860 - SettingTest::testCreateInvalidURLSettings - Fail for the right reasons
The test checks that invalid URLs return an error.  The URLs given where
invalid -- but for the wrong reason.  In unit-tests, `CIVICRM_UF_BASEURL`
defaulted to the invalid `http://FIX ME`.

After unifying `civicrm.settings.php`, the `CIVICRM_UF_BASEURL` was valid.

The intention in the test was to include a weird character (`*`) to make the
test fail, but we need a weirder/more-incorrect character.