CRM-17951 modify transaction view to show the financial account of the transaction
Previously it was showing the financial type of the contribution - but the payment could have a different financial account, although that would not be the norm and for most users no difference would be apparent. I added a group_concat in case there are multiple accounts associated with a transaction now, or in the future.

Note we have separately discussed exposing line item detail - this got stalled on how to make it look tidy in the UI