CRM-18212 allow triggers to be managed offline.
This commit adds a setting which when set (in civicrm.settings.php - not through the UI) allows you to designate the site
as having it's triggers managed offline. This is good for when it really is not appropriate to give the web use that mysql permission. This is likely the case when replication is in play because it can allow privilledge escalation