CRM-18443 add unit test & refactor code into smaller function
On starting to dig into the dedupe code I find that there are 2 different paths, one that is used by the batch merge
and one used in other cases. I'm hoping to rationalise that & add testing with the focus being on small refactors + added tests.

I'm less sure about adding an api as I work into the internals here....

Replact duplicate code with a function to calculate the cacheKey

This part of investigation / cleanup for CRM-18443