CRM-17335 reduce use of nullArray
The null array static is a work-around for pass-by-reference signatures but the executeQuery no longer requires it.

It's not actually 'just' tidy-up as the array can become contaminated, leading to hard to diagnose bugs.

Note there is one instance of a similar tidy up on contribution::create in this - the  parameter is no longer mandatory & we call this function so often I feel confident about removing the stray & too