bin/regen.sh - Drop-create table `zipcodes`
The original `regen.sh` script drops the table `zipcodes` after creating it.
That works fine in normal execution (when `regen.sh` finishes without
erorr).  But if there is an error, then the table gets left in the DB -- and
creates a conflict when you try re-run `regen.sh`.

This patch does a 'drop-create' so that subsequent runs will work despite a
previous error.