CRM_Utils_Check - Suggest using `[cms.root]`, etal
In "Admin => Resource URLs" and "Admin => Directories", sites may have a
confusing mix of absolute and relative paths, especially if they originated
in <=v4.6.  This complicates migrations (e.g.  moving to a new server or
syncing with a dev/staging system).  In 4.7+, we can use more portable path
variables.

This is a soft suggestion (NOTICE) and not a validation error or warning.
There are obscure scenarios where one would want to use absolute addresses;
but for most sites, it's an extraneous maintenance cost.