CRM-18958 add setting to control addion of limit to the dedupe screen.
We have struggled to find a compromise between the needs of larger sites not to crash when deduping & the UI experience of a 'batch' not being a batch - ie. finding duplicates for the first 1000 contacts could result in 67 or 6700. In discussion it was clear that even if we agreed on a UI to make this work the default for a large site (e.g 1000 to prevent crashes) was different to what would work for a small site - so I concluded we would need a setting. Given that it seemed to make sense to focus on adding the setting, which would provide the immediate protection needed for larger sites without compromising UI on small sites, and to put the time into that rather than solving the now less important batchin issue.

Note I did some tidy up on the settings form code too...