CRM_Utils_QueryFormatterTest - Check query results (when possible)
This test class is designed to be somewhat useful even when the underlying
search engine (InnoDB FTS, Solr) is unavailable.

However, the SQL LIKE engine is always available, and InnoDB FTS is
sometimes available.  In these cases, execute a query and see if it produces
the expected results.

See also: NYSS #9692