CRM-18513 - WordReplacements - Load en_US when locale is unspecified
On a new installation, the value in `CRM_Core_I18n::$locale` is empty.  The
word-replacement index stored in `civicrm_domain.locale_custom_strings` is
indexed by locale (e.g.  `array('en_US' => ...replacements...)`).  Thus, it
fails to locate the word-replacements (even if there are
word-replacements in the *effective* locale, `en_US`).

With this patch, the word-replacements explicitly fallback to checking `en_US`.