Test fix for intermittant fail.
The scheduled reminder test fails 'quite a bit', intermittently.

The set up is designed to run '1 hour before' the start_date but since the start date is measured in days not hours this is mostly
about trying to cope with timezones & stuff.

Trying to adjust the test to be based on day rather than an hour to
try to make the test more consistent. This does not affect the
thing the test is intended to test