CRM-19076 add test for hooks.
In resolving how to deal with addresses for this ticket we decided to err on the side of 'never lose data' ie. doing things like keeping 2 identical addresses if they have different location types. THe justification was that less conservative logic could be implemented using tests. This test adds the logic for use case where address data is more aggressively updated.

In doing so I felt the existing hook was almost unworkable & opted
to add a different hook, specific to addesses.