CRM-19074 attempt to replicate the bug in a test
I didn't actually succeed in replicating it - although I increased coverage.

The issue is that when there is no consistent differentiation
between a custom checkbox field which is no set vs one that
has been set to empty(unchecked).

There is an inconsistent data storage difference whereby it is sometimes stored as an empty string & sometimes as NULL but this does not seem to be something
that is deliberate.