CRM-19068 clean up Group Nesting class.
As identified in the ticke the add function is using non-standard
handling,
as, it turns out is the remove.

However, on further delving I found a lot of functions in this class that
do not appear to be in use. In addition the whole groupIterator complexity
seems to be unused.