CRM-19130 - CRM_Core_BAO_Cache - Do not rely on DAO::fields()
For CRM-19130, we want all calls to `fields()` to be hookable, which means
that `fields()` can only run in a bootstrapped system (where hooks are
reliably working).

This should be *mostly* OK.  However, during bootstrap, we rely on the SQL
cache which means that `CRM_Core_BAO_Cache` must be operational -- even if
`fields()` cannot be used.  Consequently, the DAO-based query-builder cannot
be used.