CRM-17789 - CRM_Core_Smarty - Remove constructor override
This is unnecessary and is part of a cyclical dependence among PHP 7 PRs.
With this resolved, we should be able to resolve some of the other PR's
in a reasonable order.

This is an adaptation of one part of #8455.  It *should* enable merging or
testing https://github.com/civicrm/civicrm-packages/pull/150 without #8455.