CRM-19130 - Add test for regression in CRM_Core_BAO_Cache
The earlier patch for CRM-19130 broke support for reading individual items
from the SQL cache.  Although the test-suite hits the cache layer a lot
(e.g.  CRM_Utils_Cache_SqlGroupTest), the bug wasn't exposed by other tests
because (a) caches are replicated in-memory and in SQL (b) caches are often
read in batch mode.

This explicitly flushes the in-memory cache and reads the SQL cache.