CRM-19223 - bin/encryptDB.php - Completely disable by default
Arguably, it's sufficient to check for administrative privileges...  since
an administrator can do this kind of damage anyway...  but this use-case is
so broad, dangerous, and uncommon that it really shouldn't be available
unless you take extraordinary measures.

It's tempting to remove the file entirely, but some upgrade workflows don't
properly handle deleted files, and some users may still want access to this
code so that they can run it themselves.