CRM_Utils_REST - Always boot CMS, even in error conditions
Now-a-days, HTTP headers are outputted using CMS functions, which means
that the CMS must always boot -- even if you're only displaying an error
message. This caused regressions in the REST error handling.  Why?
The boot sequence is really quirky with the REST calls... and the
required CMS functions were not always available.

This revision ensures that (one way or another) the CMS does get booted.
Notably, though, in the error scenarios, it boots *without* loading or
authenticating a user.