CRM-15601, CRM-19027 - Fix MyISAM DB check
3fa9688a5 introduced a subtle break in how it skips temp tables:

 * (Old) `strpos($dao->Name, '_temp') !== FALSE`
 * (New) `TABLE_NAME NOT LIKE '%temp'`

This is relevant when the table name is appended with a randomized suffix.

I don't the performance issue being addressed by CRM-19027.  If there's an
issue with this hurting indexes, then maybe the check should be done in PHP
again? Alternatively, maybe we should patch all the code which makes temp
tables to enforce `ENGINE=InnoDB`?