Fixed: CRM-19360
The import parser for contacts tried and failed to compare grettings because the exported data already had tokens replaced with data.  On import, the tokens would not be found and would fail.  I implemented
a conversion between the token and available data from the imported row to generate a regex to validate the greeting value.