CRM-20513 Fix recur status update & et next recur date based on payment date not current date.
The code to set a recurring sequence to Completed when all contributions received turned out
to be flawed. This fixes that.

This commit extends the test on how dates are being extended. I didn't replicate the problem
I was expecting to see but it makes sense that recurring transactions could be imported on
a different day to the current one and that date is the important one