Fix line item fetch on membership template with priceset.
In conjunction with checking Pradeep's code tidy up I found that the code here is not currently presenting membership
details correctly. I used Pradeep's solution, but I also clarified the process to process quickConfig as close to
the place where it is used as possible - in this case it directly affects the display of the lines.

Ideally we would assign something clearer to the template such as displayLineItemRows rahter than these mystical
parameters"
error: pathspec 'Fix line item fetch on membership template with priceset.

In conjunction with checking Pradeep's code tidy up I found that the code here is not currently presenting membership
details correctly. I used Pradeep's solution, but I also clarified the process to process quickConfig as close to
the place where it is used as possible - in this case it directly affects the display of the lines.

Ideally we would assign something clearer to the template such as displayLineItemRows rahter than these mystical
parameters