CRM-19394: Fix relative dates in “smart groups”.
When using a relative date, `member_join_date_relative` should be
converted into two a `_high` and a `_low` variables. But this
operation didn’t work correctly and only the `_high` value was kept in
the final SQL query.

The problem comes from `CRM_Contact_BAO_Query::convertFormValues()`,
when `CRM_Contact_BAO_Query::fixDateValues()` is called and to update
the content of `$formValues`. But since the main “foreach” loop in
`CRM_Contact_BAO_Query::convertFormValues()` already iterated over the
`_low` value (which was empty), this value will be missed.

The fix is to change `CRM_Contact_BAO_SavedSearch::getFormValues()`
and to push the `_relative_date` at the top of the array that will be
treated by `CRM_Contact_BAO_Query::convertFormValues()`.

(More information is available in CRM-19625).