CRM_Utils_HttpClientTest - Use special purpose hostnames
This test issues a few real HTTP requests which are supposed to succeed or
fail in a prescribed way.  However, the examples were chosen for expedience
and leached off other/unrelated sites...  but this tends to break from
time-to-time (when the example system goes to HTTPS-only or changes their
content).

With this patch, all the examples point to dedicated sites which are
purpose-built for supporting integration tests.