Ensure presence of a display name in ActionSchedule.
See CRM-19786. ActionSchedule::sendReminderEmail() was presuming the presence of a display_name token, but this was not guaranteed without notifying TokenProcessor that such a token should be populated. Since we aren't using a token to set the email recipient name, this modification simply looks up the display_name for use via Contact,get API.