CRM-19801: Fixing changing a relationship type from A to B or vice versa
If you tried to change (A) relationship type of an exiting relationship (Like parent of) to another (B) relationship type (Like Child of) or vice versa then
it will show you that the relationship get updated but it will not be it actually get updated.

The issue is related to how postProcess handle contact_id_a & contact_id_b paramterters .

So in this PR I've refactored the postProcess method for the relationship form and with this refactoring I've have fixed this issue too
by properly hanlding conact_id_a and contact_id_b parameters.