CRM-19824 - Container::loadContainer - Always compile listeners
In the typical (cache-enabled) flow, the container is compiled to identify
any services tagged with `kernel.event_subscriber`. However, if you disable
caching, then it skipped compilation, and it failed to identify all the
listeners.

With this patch, we always detect the listeners -- even if caching is
disabled.

(This issue identified while working on tests for CRM-19690.