Fixed invalid unit test blocking merge of fix for CRM-19828.
Test "should update the UI after changing the model" was failing (even without the patch
for CRM-19828) depending on the timezone in which the test was run. PhantomJS (at least
of as v1.9.8) doesn't work too well with timezones. A Date object created from a string in
the format YYYY-MM-DD assumes a time of midnight and a GMT timezone, which makes valid
comparisons between widget input and output impossible for those east of GMT.