CRM-19973 Set settings defaults during reset function.
I used buildkit to generate & run the extension test. The issue was that the settings
were being loaded but later the \Civi::reset(); function was being called - From the testListener I think
and it was flushing the settings bag so it no longer held default variables

I did look at putting this in the boot function instead and only rejected that for fear of unintended
consequences.