CRM-19934 improvement on acl cache.
This change means that when we are checking the allow function we do not add the harmful
OR to the query.

The 'view my contact' & 'edit my contact' permissions have already been checked for.

Adding them into the ACL query with an OR causes serious performance issues, as
elsewhere commented. We can & should avoid it here