CRM-20036 Don't retrieve all soft contributions if not needed.
This is an alternate fix to https://github.com/civicrm/civicrm-core/pull/9849 which basically removes the entire function. There appears to be only one place where it is called without id being set so we should simply require id & just handle that. In that one place we can use the api to do what the function was doing