CRM-20193 show financial_trxn currency on batch form.
To replicate this go to civicrm/financial/batch?reset=1&action=add & add a new batch arriving at
civicrm/batchtransaction?reset=1&bid=2

The table at the bottom of the page is affected.

This is a portion of the commit by pradpnayak which I understand, was able to test and can approve the code for.

There is still a portion on the ticket I have not yet understood enough to approve