CRM-20299 - CiviMail - "Location Type" requirement feels wonky
If you select a "Location Type" and then deselect it, the field changes
to "Automatic". This is the default value (which is acceptable), but the UI
flags it as erroneous.

The "Selection Method" also defaults to "Automatic".  These fields are oddly
a bit different (e.g.  "Automatic" translates to `location_type_id==null`
and `email_selection_method=automatic`), so the fields are a little
different, but this revision brings them closer.