CRM-19385 revert changes to order by id on prevnext_cache.
After PR #10019 was merged @bhahumanists reported on #9801 that smart group ordering was broken. The
report is lacking detail and I could not replicate it, but am reverting the search changes for safety.

After this revert sites can still experiment with dropping/altering the primary key on
-civicrm_cache
-civicrm_group_contact_cache
-civicrm_acl_cache

But not civicrm_prevnext_cache.

Although I could not replicate the search issue it seems to me that our Ubuntu test suite
is failing where the other one is not on search-related items, suggesting that
for some reason Ubuntu does not default to the natural search order, due to
some config or other difference