CRM-20264, Added Card type and Pan Truncation form fields on Payment form
I deliberately excluded card_type because I think we need to review the appropriateness of
the field first. My issue is that the wrangling wound up really wierd & it was apparent the issue
is that card_type is 'Visa' whereas card_type_id is 1 - we should either rename the field
or store the word.