GenCode - Reduce merge-conflicts on AllCoreTables
The previous checksum was based on the full schema spec.  This is
technically safe, but many irrelevant schema changes (such as editing field
comments) will ding the checksum.

The main goal here is to avoid `$template->run()` (line ~44) and its
code-beautifier (which is expensive/slow).  This revised algorithm for
`needsUpdate()` should:

 * Still be "fast enough" (given that it's a singular file built from a simple template)
 * Still be "accurate enough" (given that `getRaw()` is a fair predictor of `run()`)
 * Be less prone to merge-conflics (given that it doesn't use the checksum)