CRM-20453 - XML schema clean up: <value> tag
The <value> tag only appeared once in all the schema defintion files thus
it seemed to me to be a typo of the (no longer used) <values> tag.

Also, I searched CRM/Core/CodeGen/Specification.php for ["']value["'] and
found no occurences which makes me think this tag is not parsed.