CRM-19821 set activity_status_id to 'export'
Export is a kind of weird & wonderful setting that is undocumented. It affects
whether fields are available by default in export & search, sometimes.

Generally I would say all 'user facing fields' should be set to export or somewhere
or other we will wind up hacking around it. I recently removed a lot
of code for payment_instrument_id that was just compensating for this var being
false rather than true.