CRM-20481 temporarily switch back while we assess / fix performance.
We have been seeing performance issues with the use of the api for the dashboard. This is a back-out so
we can figure that out before re-adding. Note this should not be merged to master as there are
commits that 'add up to this' in master. There has been some git mistake this round :-(

I'll figure out how to exclude