CRM-20316 - Fix "New Mailing" regression. `subject` should not be required.
When you navigate to "New Mailing" (`civicrm/a/#/mailing/new`) in the UI, it
crashes on a white-screen.  Inspecting the console shows that the call to
`Mailing.create` is failing because `subject` is missing.  But at this point
we're only creating a *draft* mailing -- so you don't have or need a
fully-formed mailing yet.