CRM-20316 - Fix "New Mailing" regression. Update tests.
`testWithoutParam_create()` asserts that *generally* you need to provide
some params to create a record.  This was true for Mailings in <=4.7.18
because `create_id` was required.  The main goal of a98504a72 was to make
`created_id` optional, which actually meant that *no data* would be
required (making `testWithoutParam_create()` non-applicable/inaccurate.

(This was obscured at the time because a98504a72 simultaneously made
`subject` mandatory, but that led to the regression in "New Mailing".  Since
neither `created_id` nor `subject` is required...  there's nothing left to
be required.  The correct behavior is to skip `testWithoutParam_create()`
for mailings.)