CRM-20397 - crmMailingRadioDate - Less spammy
The crmUiAlert can be pretty spammy -- e.g.  if you arrow up/down to select
the delivery time (and if your time is to close the threshold for 'past'),
then the UI spits out lots of redundant dialogs.  Once you choose a valid
time, it still displays the old alert.

With this patch, only display one alert.  And close it when you've got a
good date-time.