CRM-20600 - Civi\Angular\Manager::getStrings() - Scan raw HTML strings instead of *.html files
Rationale: If the raw HTML can be modified via hook, then the HTML files are
not truly representative of the final output.  (e.g.  a hook might add a new
string to some HTML.)

Note: This patch means that we no longer benefit from the cache in
`CRM_Core_Resources_Strings`.  However, the typical user still pulls data
from a cache because we're using `AssetBuilder` to manage
`angular-modules.json`.

Note: `getStrings()` now requires a call to `getPartials()`, which is
expected to become more expensive (firing a hook).  To avoid building the
partials twice, we introduce a private cache.