CRM-20600 - civicrm/a - Avoid dirty caches for permissioned modules
The default behavior is to include a `bundleUrl` for a bundle of "all
Angular modules".  However, some Angular modules like `crmMailing` are only
displayed if the current user has permission to access them.  This means
that the list of "all Angular modules" can be different (depending on who
you are; albeit in practice there's a small# of permutations).

This patch ensures that two users who access a different list of modules
will see a different `bundleUrl`.