CRM-20756 - Add Angular `ui.bootstrap` library
This adds the library
[ui.bootstrap](https://angular-ui.github.io/bootstrap/), which makes it
easier to include the functional elements of Bootstrap in AngularJS pages.

In the past, adding this was problematic, but now the situation has
improved:

1.  civicrm-core has upgraded to AngularJS 1.5.  (In the past, civicrm-core
used AngularJS 1.3, which was incompatible.)

2.  civicrm-core allows optional AngularJS modules.  (In the past, all
modules were loaded unconditionally, which could result in loading
new/unnecessary code for legacy users.  Now, you can opt-out of this
behavior with `basePages=>array()`.)