CRM-20847 Support custom api with composite primary keys
Do not add id if the field does not exist and add unit tests.

This adds a couple of checks in the api to not add the id field
Note that Grant is an example of an api which uses a uniqueName on
id & I checked manually that is still returned. SyntaxConformance
tests cover this a lot, but I still also added one check to the grant test.