CRM-20600 - Upgrader - Display warning if asset-caching is likely to fail
After releasing v4.7.21, there were several inquiries about sites in which
Angular pages failed to load.  In cases diagnosed so far, this has been a
result of inconsistent configurations or overzealous HTTP restrictions
applied to the `imageUploadDir`.

Ordinarily, we would address this kind of thing with a status check.
Unfortunately, the problem breaks that "Status" page...  so we have to
display a bit sooner.