regen.sh - Fix stale function call in sql/GenerateData.php
In [#10463](https://github.com/civicrm/civicrm-core/pull/10463), the
function signature of `getInstrumentFinancialAccount()` changed slightly -- in that
the parameter became mandatory.

Grepping for other calls to `getInstrumentFinancialAccount()` shows that one of the unit-tests
works with example input `4`, so this does the same.

Before
======

Running "regen.sh" fails -- because `getInstrumentFinancialAccount` is called without a parameter.

After
=====

Running "regen.sh" completes -- because `getInstrumentFinancialAccount` is called with `4`.

Acceptance Prompts
==================

 * The example data produced by `regen.sh` should contain suitable/similar values for financial records.
 * Determine whether `4` is a suitable value.