CRM-20821 - Use smarter logic to adjust img URL
This change is the crux of CRM-20821.

Before this change, editing a premium product would automatically
break the image URL. After this change, premium products can be edited
while retaining the image URL. Functionality is preserved which changes
the URL to a local URL when possible.

This commit adds two new Utils functions (along with tests) to handle
the logic of changing a supplied URL to a local URL.