CRM-20759 (follow up fix) Primary address is 'distributed' over several ID
This wasn't picked up originally in the test as it was only using one field of the primary type.

The old code was indexing by 'addressCnt' (checking that the location type id was different each time). I
have switched to index by location_type_id - we should only have one address per location type per contact.

I also removed the code that sets 'is_primary' if we are importing the first address for a contact.
That has been handled by the BAO for some time