CRM-20969 alternate approach to resolving php 7.1 errors
I am proposing this as an alternate approach for fixing the 7.1 errors.

Note this only covers one report - I think we should move that function which is actually
generic to the parent class, however I think the one in contribute.summary is
a bit more nuanced.

Note this is also the approach taken by the extendedreport
extension (which takes it a little further). I will remove the declaration of
groupByArray from that extension if we add it here

I do have a medium term plan to move a lot of the helpers from extended report back into core.

Change-Id: I377d33a7417ff93169f12ccbda77f3662da5bac1