CRM-20926 - Resolve "exception"/"exceptions" discrepency. Improve test.
The PHPIDS config format uses the property:

```
exceptions => array('field_1)`
```

But our XML format uses the tag:

```
<exception>field_1</exception>
```

Grammatically, those both make sense. But they were getting confused.

This modifies the XML parser to map the tags `<exception>` to property
`exceptions`, and it updates the unit-test to ensure that the data
(from the route and the default config) are correctly merged.